Fix tests -- breaks if SimpleObject imported in multiple test files
Not going to figure this out now. Probably needs a cleanup fixture.
